Fixed Posix time and utility comp on posix

+++ src/Utilities.d
@@ -7,6 +7,6 @@ module Utilities;
 string MemString(ubyte* mem, uint pos) pure
 {
     import core.stdc.string : strlen;
-    uint len = strlen(cast(char*)&mem[pos]);
+    size_t len = strlen(cast(char*)&mem[pos]);
     return cast(string)mem[pos..pos+len];
 }
